+++ esp_video/src/device/esp_video_isp_device.c
@@ -1984,10 +1984,11 @@ esp_err_t esp_video_isp_start_by_csi(const esp_video_csi_state_t *state, const s
         ESP_GOTO_ON_ERROR(esp_isp_enable(isp_video->isp_proc), fail_2, TAG, "failed to enable ISP");
 
 #if CONFIG_ESP_VIDEO_ENABLE_ISP_VIDEO_DEVICE
-        if (COLOR_SPACE_TYPE(isp_in_color) == COLOR_SPACE_RAW) {
-            isp_video->af_support = 0;
-        } else {
+        if ((COLOR_SPACE_TYPE(isp_in_color) == COLOR_SPACE_RAW) &&
+                (COLOR_SPACE_TYPE(isp_out_color) != COLOR_SPACE_RAW)) {
             isp_video->af_support = 1;
+        } else {
+            isp_video->af_support = 0;
         }
 
         META_VIDEO_SET_FORMAT(isp_video->video, width, height, V4L2_META_FMT_ESP_ISP_STATS);