+++ esp_video/src/esp_video_cam.c
@@ -5,96 +5,133 @@
  */
 
 #include "esp_log.h"
+#include "esp_cam_sensor.h"
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+#include "esp_cam_motor.h"
+#endif
 #include "esp_video_ioctl.h"
-#include "esp_video_sensor.h"
+#include "esp_video_cam.h"
 
 struct control_map {
-    uint32_t esp_cam_sensor_id;
+    uint32_t esp_cam_priv_id;
     uint32_t v4l2_id;
 };
 
-static const char *TAG = "esp_video_sensor";
+typedef enum cam_dev_type {
+    CAM_DEV_SENSOR = 0,
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+    CAM_DEV_MOTOR
+#endif
+} cam_dev_type_t;
+
+static const char *TAG = "esp_video_cam";
 
 /**
  * Todo: AEG-1094
  */
-static const struct control_map s_control_map_table[] = {
+static const struct control_map s_sensor_control_map_table[] = {
     {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_JPEG_QUALITY,
-        .v4l2_id = V4L2_CID_JPEG_COMPRESSION_QUALITY,
-    },
-    {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_3A_LOCK,
-        .v4l2_id = V4L2_CID_3A_LOCK,
-    },
-    {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_FLASH_LED,
-        .v4l2_id = V4L2_CID_FLASH_LED_MODE,
-    },
-    {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_VFLIP,
-        .v4l2_id = V4L2_CID_VFLIP,
-    },
-    {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_HMIRROR,
-        .v4l2_id = V4L2_CID_HFLIP,
-    },
-    {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_GAIN,
+        .esp_cam_priv_id = ESP_CAM_SENSOR_GAIN,
         .v4l2_id = V4L2_CID_GAIN,
     },
     {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_EXPOSURE_US,
+        .esp_cam_priv_id = ESP_CAM_SENSOR_EXPOSURE_US,
         .v4l2_id = V4L2_CID_EXPOSURE_ABSOLUTE,
     },
     {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_EXPOSURE_VAL,
-        .v4l2_id = V4L2_CID_EXPOSURE,
-    },
-    {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_STATS,
+        .esp_cam_priv_id = ESP_CAM_SENSOR_STATS,
         .v4l2_id = V4L2_CID_CAMERA_STATS
     },
     {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_AE_LEVEL,
+        .esp_cam_priv_id = ESP_CAM_SENSOR_AE_LEVEL,
         .v4l2_id = V4L2_CID_CAMERA_AE_LEVEL
     },
     {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_GROUP_EXP_GAIN,
+        .esp_cam_priv_id = ESP_CAM_SENSOR_GROUP_EXP_GAIN,
         .v4l2_id = V4L2_CID_CAMERA_GROUP
-    }
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_SENSOR_EXPOSURE_VAL,
+        .v4l2_id = V4L2_CID_EXPOSURE,
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_SENSOR_JPEG_QUALITY,
+        .v4l2_id = V4L2_CID_JPEG_COMPRESSION_QUALITY,
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_SENSOR_3A_LOCK,
+        .v4l2_id = V4L2_CID_3A_LOCK,
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_SENSOR_FLASH_LED,
+        .v4l2_id = V4L2_CID_FLASH_LED_MODE,
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_SENSOR_VFLIP,
+        .v4l2_id = V4L2_CID_VFLIP,
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_SENSOR_HMIRROR,
+        .v4l2_id = V4L2_CID_HFLIP,
+    },
 };
 
-static const struct control_map s_control_ioctl_table[] = {
+static const struct control_map s_sensor_control_ioctl_table[] = {
     {
-        .esp_cam_sensor_id = ESP_CAM_SENSOR_IOC_S_TEST_PATTERN,
+        .esp_cam_priv_id = ESP_CAM_SENSOR_IOC_S_TEST_PATTERN,
         .v4l2_id = V4L2_CID_TEST_PATTERN,
     },
 };
+
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+static const struct control_map s_motor_control_map_table[] = {
+    {
+        .esp_cam_priv_id = ESP_CAM_MOTOR_POSITION_CODE,
+        .v4l2_id = V4L2_CID_FOCUS_ABSOLUTE,
+    },
+    {
+        .esp_cam_priv_id = ESP_CAM_MOTOR_MOVING_START_TIME,
+        .v4l2_id = V4L2_CID_MOTOR_START_TIME,
+    }
+};
+#endif
 
 /**
  * @brief Get control ID map pointer based on V4L2 control ID
  *
  * @param v4l2_id V4L2 control ID
- * @param ioctl   Sensor command type pointe, true: it is sensor ioctl command; false: it is sensor parameter ID command
+ * @param ioctl   Camera device command type pointe, true: it is sensor ioctl command; false: it is sensor parameter ID command
+ * @param type    Camera device type pointer
  *
  * @return
  *      - Control ID map pointer on success
  *      - NULL if failed
  */
-static const struct control_map *get_v4l2_ext_control_map(uint32_t v4l2_id, bool *ioctl)
+static const struct control_map *get_v4l2_ext_control_map(uint32_t v4l2_id, bool *ioctl, cam_dev_type_t *type)
 {
-    for (int i = 0; i < ARRAY_SIZE(s_control_map_table); i++) {
-        if (s_control_map_table[i].v4l2_id == v4l2_id) {
+    for (int i = 0; i < ARRAY_SIZE(s_sensor_control_map_table); i++) {
+        if (s_sensor_control_map_table[i].v4l2_id == v4l2_id) {
             *ioctl = false;
-            return &s_control_map_table[i];
+            *type = CAM_DEV_SENSOR;
+            return &s_sensor_control_map_table[i];
         }
     }
 
-    for (int i = 0; i < ARRAY_SIZE(s_control_ioctl_table); i++) {
-        if (s_control_ioctl_table[i].v4l2_id == v4l2_id) {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+    for (int i = 0; i < ARRAY_SIZE(s_motor_control_map_table); i++) {
+        if (s_motor_control_map_table[i].v4l2_id == v4l2_id) {
+            *ioctl = false;
+            *type = CAM_DEV_MOTOR;
+            return &s_motor_control_map_table[i];
+        }
+    }
+#endif
+
+    for (int i = 0; i < ARRAY_SIZE(s_sensor_control_ioctl_table); i++) {
+        if (s_sensor_control_ioctl_table[i].v4l2_id == v4l2_id) {
             *ioctl = true;
-            return &s_control_ioctl_table[i];
+            *type = CAM_DEV_SENSOR;
+            return &s_sensor_control_ioctl_table[i];
         }
     }
 
@@ -104,34 +141,48 @@ static const struct control_map *get_v4l2_ext_control_map(uint32_t v4l2_id, bool
 /**
  * @brief Get control internal operation parameters
  *
- * @param cam_dev  Camera sensor device pointer
+ * @param cam      Camera device pointer
  * @param ctrl     V4L2 external control pointer
  * @param qdesc    Camera sensor device query description pointer
  * @param buf_ptr  Actual data buffer pointer
  * @param buf_size Actual size buffer pointer
  * @param ioctl    Sensor command type pointe, true: it is sensor ioctl command; false: it is sensor parameter ID command
+ * @param type     Camera device type pointer
  *
  * @return
  *      - ESP_OK on success
  *      - Others if failed
  */
-static esp_err_t get_opt_value_desc(esp_cam_sensor_device_t *cam_dev, struct v4l2_ext_control *ctrl,
-                                    esp_cam_sensor_param_desc_t *qdesc, void **buf_ptr, size_t *buf_size, bool *ioctl)
+static esp_err_t get_opt_value_desc(esp_video_cam_t *cam, struct v4l2_ext_control *ctrl, esp_cam_sensor_param_desc_t *qdesc,
+                                    void **buf_ptr, size_t *buf_size, bool *ioctl, cam_dev_type_t *dev_type)
 {
     esp_err_t ret;
     const struct control_map *control_map;
 
-    control_map = get_v4l2_ext_control_map(ctrl->id, ioctl);
+    control_map = get_v4l2_ext_control_map(ctrl->id, ioctl, dev_type);
     if (!control_map) {
         ESP_LOGE(TAG, "ctrl id=%" PRIx32 " is not supported", ctrl->id);
         return ESP_ERR_NOT_SUPPORTED;
     }
 
-    qdesc->id = control_map->esp_cam_sensor_id;
+    qdesc->id = control_map->esp_cam_priv_id;
     if (*ioctl == false) {
-        ret = esp_cam_sensor_query_para_desc(cam_dev, qdesc);
+        if (*dev_type == CAM_DEV_SENSOR) {
+            ret = esp_cam_sensor_query_para_desc(cam->sensor, qdesc);
+        } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+            ret = esp_cam_motor_query_para_desc(cam->motor, qdesc);
+#else
+            ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+        }
         if (ret == ESP_ERR_NOT_SUPPORTED) {
-            ESP_LOGD(TAG, "sensor %s doesn't support to query parameter description", cam_dev->name);
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+            const char *name = *dev_type == CAM_DEV_SENSOR ? cam->sensor->name : cam->motor->name;
+#else
+            const char *name = cam->sensor->name;
+#endif
+            ESP_LOGD(TAG, "sensor %s doesn't support to query parameter description", name);
 
             *buf_ptr = &ctrl->value;
             *buf_size = sizeof(ctrl->value);
@@ -175,16 +226,16 @@ static esp_err_t get_opt_value_desc(esp_cam_sensor_device_t *cam_dev, struct v4l
 }
 
 /**
- * @brief Set control value to camera sensor device
+ * @brief Set control value to camera device
  *
- * @param cam_dev  Camera sensor device pointer
+ * @param cam      Camera device pointer
  * @param controls V4L2 external controls pointer
  *
  * @return
  *      - ESP_OK on success
  *      - Others if failed
  */
-esp_err_t esp_video_set_ext_ctrls_to_sensor(esp_cam_sensor_device_t *cam_dev, const struct v4l2_ext_controls *controls)
+esp_err_t esp_video_cam_set_ext_ctrls(esp_video_cam_t *cam, const struct v4l2_ext_controls *controls)
 {
     esp_err_t ret = ESP_ERR_INVALID_ARG;
 
@@ -192,16 +243,25 @@ esp_err_t esp_video_set_ext_ctrls_to_sensor(esp_cam_sensor_device_t *cam_dev, co
         bool ioctl;
         void *value_ptr;
         size_t value_size;
+        cam_dev_type_t dev_type;
         esp_cam_sensor_param_desc_t qdesc;
         struct v4l2_ext_control *ctrl = &controls->controls[i];
 
-        ret = get_opt_value_desc(cam_dev, ctrl, &qdesc, &value_ptr, &value_size, &ioctl);
+        ret = get_opt_value_desc(cam, ctrl, &qdesc, &value_ptr, &value_size, &ioctl, &dev_type);
         if (ret != ESP_OK) {
             break;
         }
 
         if (ioctl) {
-            ret = esp_cam_sensor_ioctl(cam_dev, qdesc.id, value_ptr);
+            if (dev_type == CAM_DEV_SENSOR) {
+                ret = esp_cam_sensor_ioctl(cam->sensor, qdesc.id, value_ptr);
+            } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+                ret = esp_cam_motor_ioctl(cam->motor, qdesc.id, value_ptr);
+#else
+                ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+            }
             if (ret != ESP_OK) {
                 ESP_LOGE(TAG, "failed to set ioctl id=%" PRIx32, ctrl->id);
                 break;
@@ -248,7 +308,15 @@ esp_err_t esp_video_set_ext_ctrls_to_sensor(esp_cam_sensor_device_t *cam_dev, co
                 return ESP_ERR_NOT_SUPPORTED;
             }
 
-            ret = esp_cam_sensor_set_para_value(cam_dev, qdesc.id, value_ptr, value_size);
+            if (dev_type == CAM_DEV_SENSOR) {
+                ret = esp_cam_sensor_set_para_value(cam->sensor, qdesc.id, value_ptr, value_size);
+            } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+                ret = esp_cam_motor_set_para_value(cam->motor, qdesc.id, value_ptr, value_size);
+#else
+                ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+            }
             if (ret != ESP_OK) {
                 ESP_LOGE(TAG, "failed to set ctrl id=%" PRIx32, ctrl->id);
                 break;
@@ -260,16 +328,16 @@ esp_err_t esp_video_set_ext_ctrls_to_sensor(esp_cam_sensor_device_t *cam_dev, co
 }
 
 /**
- * @brief Get control value from camera sensor device
+ * @brief Get control value from camera device
  *
- * @param cam_dev  Camera sensor device pointer
+ * @param cam      Camera device pointer
  * @param controls V4L2 external controls pointer
  *
  * @return
  *      - ESP_OK on success
  *      - Others if failed
  */
-esp_err_t esp_video_get_ext_ctrls_from_sensor(esp_cam_sensor_device_t *cam_dev, struct v4l2_ext_controls *controls)
+esp_err_t esp_video_cam_get_ext_ctrls(esp_video_cam_t *cam, struct v4l2_ext_controls *controls)
 {
     esp_err_t ret = ESP_ERR_INVALID_ARG;
 
@@ -277,22 +345,39 @@ esp_err_t esp_video_get_ext_ctrls_from_sensor(esp_cam_sensor_device_t *cam_dev, 
         bool ioctl;
         void *value_ptr;
         size_t value_size;
+        cam_dev_type_t dev_type;
         esp_cam_sensor_param_desc_t qdesc;
         struct v4l2_ext_control *ctrl = &controls->controls[i];
 
-        ret = get_opt_value_desc(cam_dev, ctrl, &qdesc, &value_ptr, &value_size, &ioctl);
+        ret = get_opt_value_desc(cam, ctrl, &qdesc, &value_ptr, &value_size, &ioctl, &dev_type);
         if (ret != ESP_OK) {
             break;
         }
 
         if (ioctl) {
-            ret = esp_cam_sensor_ioctl(cam_dev, qdesc.id, value_ptr);
+            if (dev_type == CAM_DEV_SENSOR) {
+                ret = esp_cam_sensor_ioctl(cam->sensor, qdesc.id, value_ptr);
+            } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+                ret = esp_cam_motor_ioctl(cam->motor, qdesc.id, value_ptr);
+#else
+                ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+            }
             if (ret != ESP_OK) {
                 ESP_LOGE(TAG, "failed to get ioctl id=%" PRIx32, ctrl->id);
                 break;
             }
         } else {
-            ret = esp_cam_sensor_get_para_value(cam_dev, qdesc.id, value_ptr, value_size);
+            if (dev_type == CAM_DEV_SENSOR) {
+                ret = esp_cam_sensor_get_para_value(cam->sensor, qdesc.id, value_ptr, value_size);
+            } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+                ret = esp_cam_motor_get_para_value(cam->motor, qdesc.id, value_ptr, value_size);
+#else
+                ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+            }
             if (ret != ESP_OK) {
                 ESP_LOGE(TAG, "failed to get ctrl id=%" PRIx32, ctrl->id);
                 break;
@@ -304,30 +389,39 @@ esp_err_t esp_video_get_ext_ctrls_from_sensor(esp_cam_sensor_device_t *cam_dev, 
 }
 
 /**
- * @brief Get control description from camera sensor device
+ * @brief Get control description from camera device
  *
- * @param cam_dev  Camera sensor device pointer
+ * @param cam      Camera device pointer
  * @param qctrl    V4L2 external controls query description pointer
  *
  * @return
  *      - ESP_OK on success
  *      - Others if failed
  */
-esp_err_t esp_video_query_ext_ctrls_from_sensor(esp_cam_sensor_device_t *cam_dev, struct v4l2_query_ext_ctrl *qctrl)
+esp_err_t esp_video_cam_query_ext_ctrls(esp_video_cam_t *cam, struct v4l2_query_ext_ctrl *qctrl)
 {
     bool ioctl;
     esp_err_t ret;
+    cam_dev_type_t dev_type;
     const struct control_map *control_map;
     esp_cam_sensor_param_desc_t qdesc;
 
-    control_map = get_v4l2_ext_control_map(qctrl->id, &ioctl);
+    control_map = get_v4l2_ext_control_map(qctrl->id, &ioctl, &dev_type);
     if (!control_map) {
         ESP_LOGE(TAG, "ctrl id=%" PRIx32 " is not supported", qctrl->id);
         return ESP_ERR_NOT_SUPPORTED;
     }
 
-    qdesc.id = control_map->esp_cam_sensor_id;
-    ret = esp_cam_sensor_query_para_desc(cam_dev, &qdesc);
+    qdesc.id = control_map->esp_cam_priv_id;
+    if (dev_type == CAM_DEV_SENSOR) {
+        ret = esp_cam_sensor_query_para_desc(cam->sensor, &qdesc);
+    } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+        ret = esp_cam_motor_query_para_desc(cam->motor, &qdesc);
+#else
+        ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+    }
     if (ret != ESP_OK) {
         ESP_LOGD(TAG, "failed to query sensor id=%" PRIx32, qdesc.id);
         return ret;
@@ -386,23 +480,24 @@ esp_err_t esp_video_query_ext_ctrls_from_sensor(esp_cam_sensor_device_t *cam_dev
 }
 
 /**
- * @brief Query menu value from camera sensor device
+ * @brief Query menu value from camera device
  *
- * @param cam_dev  Camera sensor device pointer
+ * @param cam      Camera device pointer
  * @param qmenu    Menu value buffer pointer
  *
  * @return
  *      - ESP_OK on success
  *      - Others if failed
  */
-esp_err_t esp_video_query_menu_from_sensor(esp_cam_sensor_device_t *cam_dev, struct v4l2_querymenu *qmenu)
+esp_err_t esp_video_cam_query_menu(esp_video_cam_t *cam, struct v4l2_querymenu *qmenu)
 {
     bool ioctl;
     esp_err_t ret;
+    cam_dev_type_t dev_type;
     const struct control_map *control_map;
     esp_cam_sensor_param_desc_t qdesc;
 
-    control_map = get_v4l2_ext_control_map(qmenu->id, &ioctl);
+    control_map = get_v4l2_ext_control_map(qmenu->id, &ioctl, &dev_type);
     if (!control_map) {
         ESP_LOGE(TAG, "ctrl id=%" PRIx32 " is not supported", qmenu->id);
         return ESP_ERR_NOT_SUPPORTED;
@@ -413,8 +508,16 @@ esp_err_t esp_video_query_menu_from_sensor(esp_cam_sensor_device_t *cam_dev, str
         return ESP_ERR_NOT_SUPPORTED;
     }
 
-    qdesc.id = control_map->esp_cam_sensor_id;
-    ret = esp_cam_sensor_query_para_desc(cam_dev, &qdesc);
+    qdesc.id = control_map->esp_cam_priv_id;
+    if (dev_type == CAM_DEV_SENSOR) {
+        ret = esp_cam_sensor_query_para_desc(cam->sensor, &qdesc);
+    } else {
+#if CONFIG_ESP_VIDEO_ENABLE_CAMERA_MOTOR_CONTROLLER
+        ret = esp_cam_motor_query_para_desc(cam->motor, &qdesc);
+#else
+        ret = ESP_ERR_NOT_SUPPORTED;
+#endif
+    }
     if (ret != ESP_OK) {
         ESP_LOGE(TAG, "failed to query sensor id=%" PRIx32, qdesc.id);
         return ret;